handle response modes with one function
