+++ utils/response_mode_handler.go
@@ -1,6 +1,7 @@
 package utils
 
 import (
+	"errors"
 	"net/http"
 
 	"github.com/golang-jwt/jwt/v4"
@@ -15,6 +16,27 @@ type ResponseData struct {
 
 func (r *ResponseData) Valid() error {
 	return nil
+}
+
+func HandleAuthResponseMode(r *http.Request) (ResponseData, error) {
+	if r.URL.Query().Get("response") != "" {
+		return GetResponseDataFromJWT(r)
+	}
+
+	if r.URL.Query().Get("code") != "" {
+		return GetResponseDataFromQuery(r)
+	}
+
+	return ResponseData{}, errors.New("unable to determine response mode")
+}
+
+func GetResponseDataFromQuery(r *http.Request) (ResponseData, error) {
+	return ResponseData{
+		Code:             r.URL.Query().Get("code"),
+		State:            r.URL.Query().Get("state"),
+		Error:            r.URL.Query().Get("error"),
+		ErrorDescription: r.URL.Query().Get("error_description"),
+	}, nil
 }
 
 func GetResponseDataFromJWT(r *http.Request) (ResponseData, error) {
