--- utils/response_mode_parse.go
@@ -1,5 +1,0 @@
-package utils
-
-import "net/http"
-
-type CallbackURLResponseModeParser func(*http.Request) (ResponseData, error)
